+++ test/sql/uninstall.sql
@@ -10,12 +10,12 @@ create sequence emaj_emaj_demo_app_schema.not_an_emaj_seq;
 -- So drop this sequence and retry.
 drop sequence emaj_emaj_demo_app_schema.not_an_emaj_seq;
 
--- Call the uninstall psql script
-\i sql/emaj_uninstall.sql
+-- Call the emaj_drop_extension function
+select emaj.emaj_drop_extension();
 
 -- Check that the extension and the emaj schema are not known anymore
-\dx
-\dn
+\dx emaj
+\dn emaj
 
 -- Drop the extra extension to get a stable re-install test
 drop extension dblink;