+++ test/sql/uninstall_psql.sql
@@ -1,11 +1,11 @@
 -- uninstall_psql.sql : test of the E-Maj drop when installed with a psql script (i.e. not as an EXTENSION)
 --
 
--- Call the uninstall psql script
-\i sql/emaj_uninstall.sql
+-- Call the emaj_drop_extension function
+select emaj.emaj_drop_extension();
 
--- Check that the emaj schema are not known anymore
-\dn
+-- Check that the emaj schema is not known anymore
+\dn emaj
 
 -- Drop the extra extension to get a stable re-install test
 drop extension dblink;